Neural.Component

Description

This module defines components, which are parameterized functions which can be trained by backpropagation. By using the Arrow instance, small components can easily by combined to bigger, more complicated components.

Documentation

newtype Component' t a b

The type Component' t a b describes parameterized functions from a to b, where the parameters are of type t Analytic. When such components are composed, they all share the same parameters.

Constructors

Component'

Fields

runC :: a -> t Analytic -> b

data Component a b

A Component a b is a parameterized function from a to b for some collection of analytic parameters. In contrast to Component', when these components are composed, each component carries its own collection of parameters.

Constructors

forall t . (Traversable t, Applicative t) => Component

Fields

weights :: t Double

The specific parameter values.

compute :: Component' t a b

The encapsulated parameterized function.

initR :: forall m. MonadRandom m => m (t Double)

Randomly sets the parameters.

weightsLens :: Lens' (Component a b) [Double]

A Lens' to get or set the weights of a component. The shape of the parameter collection is hidden by existential quantification, so instead this lens uses simple lists.

activate :: Component a b -> a -> b

Activates a component, i.e. applies it to the specified input, using the current parameter values.

componentR :: MonadRandom m => Component a b -> m (Component a b)

Sets the parameters randomly, but keeps all other properties of a component.