Neural.Descent

Description

This module provides gradient descent for models. A special case of this is the backpropagation algorithm for neural networks.

Documentation

descent

Arguments

:: Foldable h

=> Model f g a b c
the model whose error should be decreased

-> Double
the learning rate

-> h a
a mini batch of samples

-> (Double, Model f g a b c)
returns the average sample error and the improved model

descentM

Arguments

:: (Foldable h, Monad m)

=> Double
the learning rate

-> h a
the mini batch of samples

-> ModelM f g a b c m Double
returns the average error

This is the monadic version of descent: It performs one step of gradient descent on a "mini batch" of samples and returns the average sample error.