Neural.Monad

Description

This module defines the ModelM monad for working with models. It is basically a state monad (which holds the model in its state) with the added functionality of random number generation.

Documentation

data ModelM f g a b c m x

The ModelM f g a b c m monad simplifies working with models of type Model f g a b c.

runModelM

Arguments

:: Monad m

=> Model f g a b c
initial model

-> StdGen
initial random number generator

-> ModelM f g a b c m x
the monadic expression to run

-> m (x, Model f g a b c, StdGen)
returns result, new model and new random number generator

Runs a computation in the ModelM f g a b c m monad.

evalModelM

Arguments

:: Monad m

=> Model f g a b c
initial model

-> StdGen
initial random number generator

-> ModelM f g a b c m x
the monadic expression to evaluate

-> m x
returns the result

Evaluates a computation in the ModelM a b c d e m monad.

modelM :: Monad m => b -> ModelM f g a b c m c

Monadic version of model.

errorM :: (Foldable h, Monad m) => h a -> ModelM f g a b c m Double

Monadic version of errorModel.

randomizeM :: Monad m => ModelM f g a b c m ()

Monadic version of modelR.

descentM

Arguments

:: (Foldable h, Monad m)

=> Double
the learning rate

-> h a
the mini batch of samples

-> ModelM f g a b c m Double
returns the average error

This is the monadic version of descent: It performs one step of gradient descent on a "mini batch" of samples and returns the average sample error.