Numeric.Neural.Layer

Description

This modules defines special "layer" components and convenience functions for the creation of such layers.

Documentation

type Layer i o = Component (Vector i Analytic) (Vector o Analytic)

A Layer i o is a component that maps a vector of length i to a vector of length j.

linearLayer :: forall i o. (KnownNat i, KnownNat o) => Layer i o

Creates a linear Layer, i.e. a layer that multiplies the input with a weight matrix and adds a bias to get the output.

layer :: (KnownNat i, KnownNat o) => (Analytic -> Analytic) -> Layer i o

Creates a Layer as a combination of a linear layer and a non-linear activation function.

tanhLayer :: (KnownNat i, KnownNat o) => Layer i o

This is simply layer, specialized to tanh-activation. Output values are all in the interval [0,1].

logisticLayer :: (KnownNat i, KnownNat o) => Layer i o

This is simply layer, specialized to the logistic function as activation. Output values are all in the interval [-1,1].

softmax :: (Floating a, Functor f, Foldable f) => f a -> f a

The softmax function normalizes a vector, so that all entries are in [0,1] with sum 1. This means the output entries can be interpreted as probabilities.