Utils.Stack

Description

This module defines the StackT monad transformer, which is simply a wrapped state monad whose state is a list.

Documentation

data StackT s m a

A computation of type StackT s m a has access to a stack of elements of type s.

pop :: Monad m => StackT s m (Maybe s)

Pops the top element from the stack. Returns Nothing if the stack is empty.

peek :: Monad m => StackT s m (Maybe s)

Peeks at the top element of the stack. Returns Nothing if the stack is empty.

push :: Monad m => s -> StackT s m ()

Pushes a new element onto the stack.

runStackT :: Monad m => StackT s m a -> [s] -> m (a, [s])

Runs a computation in the StackT s m monad.

evalStackT :: Monad m => StackT s m a -> [s] -> m a

Evaluates a computation in the StackT s m monad.

execStackT :: Monad m => StackT s m a -> [s] -> m [s]

Executes a computation in the StackT s m monad.

type Stack s = StackT s Identity

A pure stack monad.

runStack :: Stack s a -> [s] -> (a, [s])

Runs a computation in the Stack s monad.

evalStack :: Stack s a -> [s] -> a

Evaluates a computation in the Stack s monad.

execStack :: Stack s a -> [s] -> [s]

Executes a computation in the Stack s monad.