Utils.Traversable

Description

This module contains utility functions related to the Traversable typeclass.

Documentation

fromList :: (Applicative t, Traversable t) => [a] -> Maybe (t a)

Tries to create a traversable (which must also be applicative) from a list. If the list contains too few elements, Nothing is returned,

>>> fromList [1, 2, 3] :: Maybe (Identity Int)
Just (Identity 1)

>>> fromList [] :: Maybe (Identity Char)
Nothing

safeHead :: [a] -> Maybe a

Returns the head of a non-empty list or Nothing for the empty list.

>>> safeHead "Haskell"
Just 'H'

>>> safeHead ""
Nothing